# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Nodespace API
  version: 0.1.0
  summary: API for the Nodespace AI pipeline editor.
  description: >
    First-pass OpenAPI specification generated from the current Express route tree.
    The API is primarily a proxy/orchestration layer over third-party AI providers
    for image, video, audio, workflow generation, and LLM-assisted utilities.
servers:
  - url: http://localhost:3001
    description: Local development
...
paths:
...
  /api/generate-image:
    post:
      tags: [Images]
      summary: Generate an image
      description: Starts an image generation job using Google or Freepik-backed services.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FlexibleGenerationRequest'
      responses:
        '200':
          description: Job started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AsyncTaskResponse'
  /api/generate-kora:
    post:
      tags: [Images]
      summary: Generate a realism-focused image
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FlexibleGenerationRequest'
      responses:
        '200':
          description: Job started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AsyncTaskResponse'
  /api/status/{taskId}:
    get:
      tags: [Images]
      summary: Poll generic image-generation status
      parameters:
        - $ref: '#/components/parameters/TaskId'
        - in: query
          name: model
          schema:
            type: string
          description: Optional model hint forwarded to provider polling.
      responses:
        '200':
          description: Current task state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AsyncTaskResponse'
  /api/upload-image:
    post:
      tags: [Images]
      summary: Upload one or more images
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                images:
                  type: array
                  maxItems: 5
                  items:
                    type: string
                    format: binary
      responses:
        '200':
          description: Uploaded images converted to data URLs
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  urls:
                    type: array
                    items:
                      type: string
                  count:
                    type: integer
...
components:
  parameters:
    TaskId:
      in: path
      name: taskId
      required: true
      schema:
        type: string
      description: Provider task identifier.
...
  schemas:
    FlexibleGenerationRequest:
      type: object
      additionalProperties: true
      description: >
        Provider-specific request body. Common fields include prompt, image_url,
        images, model, aspect_ratio, resolution, duration, and other node-driven settings.
      properties:
        prompt:
          type: string
        model:
          type: string
        image_url:
          type: string
        image:
          type: string
        images:
          type: array
          items:
            type: string
    AsyncTaskResponse:
      type: object
      additionalProperties: true
      properties:
        data:
          type: object
          additionalProperties: true
          properties:
            id:
              type: string
            status:
              type: string
              examples: [PENDING, IN_PROGRESS, COMPLETED, FAILED]
            generated:
              type: array
              items:
                type: string
        error:
          oneOf:
            - type: string
            - type: object
              additionalProperties: true
...